<Insert commit title here> Nathan Changes:  - SCRODQB_TOP.VHD replaced asyncronous control inputs with a                    statemachine to sequence through control. Debugged 		   Statemachines and added conditional signals to 		   sensitivity list.
 - HODO_DC_QBTOP.vhd Added dc_status signal which flags when a
                     successful write occured. This triggers the
		     transistion from the writing stage to readback
		     (both SCROD and HODO). Debugged
                     statemachine and added conditional signals to
                     sensitivity list.

Issues addressed by changes:
 + VIO cores unavailable w/ Chipscope Pro: instead of using C.S. VIO to
                                           provide control signals, the
					   CTRL_SM on the SCROD will
					   sequence the sending and
					   readback automatically.

Comments:
There is no async reset for the system should there be an error. The
FPGAs must be reprogrammed to reset the system. This will need to be
addressed in the future, but it should not be necceesary for the initial
hardware tests.

The system was successfully testbenched. The Wave config. file can be
found in:
/Desktop/HMB/EIC-Beamtest-FW/SCROD_QBLinkProto/isim/HMB_DC_QBTOP_isim_beh.exe.sim/TB_simWV.wcfg

Plan of Action: Program SCROD and HODOSCOPE FGPAs and see if a
successful write and readback is achieved with the QBLink